feat(vote): burst celebration particles on every incoming vote
The "waiting for votes" screen used to be a static progress bar — voters on
flaky networks couldn't tell whether the session was alive. We now reuse the
existing CelebrationParticles burst from the random-pick reveal: each time
voterCount increments via the vote:cast socket event, a fresh particle
burst plays around the progress bar.

Implementation notes:
- Extracts CelebrationParticles into its own component file so VotePage and
  random-pick-modal can both consume it without duplicating the particle math.
- The burst is keyed on voterCount so React remounts the component on every
  increment and the lazy initializer regenerates the random positions.
- count={10} keeps each burst lighter than the reveal celebration (default
  18) so frequent retriggers don't feel noisy.
- Wrapping the Progress bar in a relative container lets the absolutely-
  positioned particles bloom upward from its baseline.
